Build.scala: ensure the defined scalaVersion is used throughout the build
This fix is required to untangle the Scala version "wired" to the
installed version of sbt.  For instance, this allows us to build
Scalding with sbt 0.13.0 and scalaVersion set to Scala 2.10.3 even
though sbt 0.13.0 uses Scala 2.10.2.

Without this change we run into the following error when having sbt
0.13.0 and Scala 2.10.3 installed locally, and set scalaVersion to
2.10.2 in projects/Build.scala:

    java.lang.ClassCastException: com.twitter.scalding.Opaque cannot be cast to java.lang.Comparable
      at com.twitter.scalding.IntegralComparator.compare(IntegralComparator.scala:54)

See the following related discussions:

- Fix compiler library in SBT 0.13.x with Scala != 2.10.2
  https://github.com/mpeltonen/sbt-idea/pull/253
- How does sbt choose which Scala version to use?
  http://stackoverflow.com/questions/14572148
- How to change Scala version for build definition?
  http://stackoverflow.com/questions/11948002