Add semi-general support for scalding modules within scald.rb.
Add --<short-module-name> (i.e. without leading scalding) to
scald.rb command line to include the related module jar to
the build and local run command lines.

Finally, add a new Json example which makes use of the new support.

Note that an extra copy of jackson is in hadoop jar thus
promote all scalding core and module jars to the front of
the generated CLASSPATH to ensure that the good copy
in avro jar is found first.