Enable read transformation in test to get correct fields in sourceTap
JsonLine previously did not have the transformations enabled 
during the test phase (transformInTest set to false).

Because of this problem the only way to use the JsonLine was to create
a sub-class and override the transformInTest method with a ‘true’ def
value. 

This workaround was working but had side-effects on the code-base:
- Renaming JsonLine during import and subclassing it
- Duplication of the JsonLine subclass all over the code

We do not see any valid reason why the test should not invoke
the transformation and thus not creating the valid set of fields
for the underlying data. With Json format you have to interpret the 
data otherwise the sourceFields will stay with their default values
(‘offset’, ‘line’).

The problem we saw was exactly this:
“could not select fields: ... from: [{2}:'offset', 'line’]”
(because of the transformation not applied and thus source fields not 
evaluated)