JsonLine: allow read transformation to get correct fields in sourceTap
JsonLine previously did not have the transformations enabled
during the test phase (transformInTest set to false).

Because of this the only way to use the JsonLine was to create
a sub-class and override the transformInTest method with a ‘true’ def
value: as JsonLine can be effectively used with JobTest only
with transformation enabled, sub-classing and overriding was 
quite common.

When using JsonLine without conversation enabled you get a 
Cascading mapping exception:
“could not select fields: ... from: [{2}:'offset', 'line’]”

Default setting for transformInTest is kept to false for backward
compatibility with 0.9.x and 0.10.x.