Set Cascading 2.6 tracing boundary
This commit set's a tracing boundary so that DocumentServices such as
Driven don't show "RichPipe" as being the originator of Cascading nodes.
Instead user code should now be the origin. The change to existing
Scalding code is kept minimal: the Job class calls an init method. If
other entry points are common it's easy enough to add them. The call
into Cascading is implemented via reflection so users can continue to
use older versions of Cascading without breakage.