Improve type safety of internal `Boxed` logic
I've been applying `wartremover`[1] to the `scalding` code base to try to
improve type safety.  One of the first things singled out was the `Boxed` logic,
for three reasons:

1. The use of `Any`
2. The use of `asInstanceOf`
3. The exception when exhausting the finite supply of `Boxed{n}` classes

I was able to fix (1), which required factoring out the functions into
object-level `def`s, since AFAICT Scala does not permit polymorphic anonymous
functions.

I was not able to fix (2) because the `Class` class is invariant in its
type parameter

I didn't understand the motivation behind the `Boxed` work-around to assess
whether or not the exception could be eliminated through more type safety.  All
I did was improve the error message to be more helpful.

1: https://github.com/puffnfresh/wartremover/tree/latest-release/core/src/main/scala/wartremover/warts