Add andThen to Mappable (#1656)
Mappable.andThen is defined in TypedSource and when invoked it
returns a TypedSource. In this patch I'm overriding that method
so that when you invoke a Mappable.andThen it returns a Mappable
object. This can be useful when the user of Mappable wants to do
toIeratable after applying a transformation using andThen