Add a setting to skip null counters
We have observed NPEs for null counters at least two companies
using scalding. We have not root-caused this issue or found
a good fix, but previously set to ignore all null counters.

Since some people rely on counters this is not a great plan,
so instead this patch makes ignoring an opt-in behavior.