Fix `planTypedWrites` when typed write happens in typed sink implementation (#1885)
During the release process in Twitter we found a customer who had a `TypedSink` implementation which did a casting of cascading's `Pipe` back to `TypedPipe` in `writeFrom` and doing typed `write` on it. This was working with 0.17.x branch but it's broken in 0.18.x since we do `planTypedWrites` only once.

This PR adds test for that and fixes it.
Even this type of `Sink`s implementation seems wrong it's better to keep scalding backward compatible with currently released version.