Alternative implementation to DeprecatedParquetInputFormat with fix (#1937)
When combining N Parquet files, the first record of files 2 to N gets skipped while the last record from the previous file is returned instead. This means losing some records while others get duplicated, quite bad.

This was fixed a month ago in apache/parquet-mr#844 but we would need to update the dependencies.

Should we do this approach or work towards updating deps?