Beam backend: use TypedPipe descriptions as names for PTransforms
Problem: When debugging Beam jobs running on Dataflow, it is hard to map a failure in the Dataflow UI back to the corresponding line in the Scalding code.

Solution: Assign the description of the TypedPipe as the name of the PTransform(this name appears in the UI). This allows mapping the error back to the
line in the Scalding code since by default, the descriptions contain the line numbers.